Nuevo Sistema de Identificacion para Clientes, más ventanas con estilo, implementacion para "atrapar errores" en Usuarios.
-- ARREGLAR --

- TABLEVIEW DE VENTASLOCALES NO FUNCIONA
- VENTAS NACIONALES SIN EMPEZAR
- HISTORIAL SIN EMPEZAR
- CAJA SIN EMPEZAR
- MÚLTIPLES VISTAS SIN ESTILO

+++ src/main/java/com/alilopez/application/controllers/NewVentaLocalController.java
@@ -35,7 +35,7 @@ public class NewVentaLocalController {
     @FXML
     void onClickSaveButton(MouseEvent event) {
         float descuento;
-        float cantidad = Integer.parseInt(cantidadTxt.getText());
+        double cantidad = Integer.parseInt(cantidadTxt.getText());
         if (descuentoTxt.getText().isEmpty()){
             descuento = 0;
         } else {
